extracting PointsetMesh from mesh.h,cpp
