+++ src/ringmesh/mesh/CMakeLists.txt
@@ -38,6 +38,7 @@ target_sources(${target_name}
         "${lib_source_dir}/mesh_builder.cpp"
         "${lib_source_dir}/mesh_index.cpp"
         "${lib_source_dir}/mesh_base.cpp"        
+        "${lib_source_dir}/pointsetmesh.cpp"        
         "${lib_source_dir}/mesh.cpp"
     PRIVATE # Could be PUBLIC from CMake 3.3
         "${lib_include_dir}/common.h"
@@ -45,6 +46,7 @@ target_sources(${target_name}
         "${lib_include_dir}/mesh_builder.h"
         "${lib_include_dir}/mesh_index.h"
         "${lib_include_dir}/mesh_base.h"
+        "${lib_include_dir}/pointsetmesh.h"
         "${lib_include_dir}/mesh.h"
 )
